+++ praktikumsplaner-frontend/frontend/src/api/MailService.ts
@@ -15,7 +15,6 @@ export default {
         )
             .then((response) => {
                 if (response.ok) {
-                    // Prüfen, ob der Statuscode OK (200) ist
                     useSnackbarStore().showMessage({
                         message: "☑ Alle Mails wurden erfolgreich versendet.",
                         level: Levels.SUCCESS,
@@ -27,7 +26,7 @@ export default {
                             message: "Fehler beim Versenden einiger Mails.",
                             level: Levels.ERROR,
                         });
-                        return praktikumsplaetze; // Liste der betroffenen Praktikumsplätze zurückgeben
+                        return praktikumsplaetze;
                     });
                 }
             })
@@ -36,7 +35,7 @@ export default {
                     message: err.message,
                     level: Levels.ERROR,
                 });
-                return Promise.reject(err); // Fehler zurückgeben
+                return Promise.reject(err);
             });
     },
 };
