+++ praktikumsplaner-frontend/frontend/src/components/common/InitialsAvatar.vue
